refactor WeekScreen
