+++ src/screens/index.js
@@ -1,5 +1,5 @@
 export { default as TodayScreen } from './TodayScreen'
-export { default as WeekScreen } from './WeekScreen/WeekScreen'
+export { default as WeekScreen } from './WeekScreen'
 export { default as ActivityDetailScreen} from './ActivityDetailScreen/ActivityDetailScreen'
 export { default as GoalsScreen } from './GoalsScreen/GoalsScreen'
 export { default as GoalScreen} from './GoalScreen/GoalScreen'
